django-ecommerce-api/requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.4
Django==4.0.4
django-allauth==0.50.0
django-cors-headers==3.13.0
django-countries==7.3.2
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.24.1
google-api-core==2.8.2
google-auth==2.8.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.3
gunicorn==20.1.0
httplib2==0.20.4
humanize==4.4.0
idna==3.3
inflection==0.5.1
jsonschema==4.16.0
kombu==5.2.4
oauthlib==3.2.0
packaging==21.3
phonenumbers==8.12.47
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==4.21.1
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
redis==4.3.4
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
stripe==4.1.0
toml==0.10.2
tornado==6.2
twilio==7.9.3
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
pre-commit==3.4.0
